+++ src/apis/axiosInstance.ts
@@ -1,4 +1,4 @@
-import axios, {AxiosError, AxiosResponse} from 'axios';
+import axios, {AxiosError, AxiosRequestConfig, AxiosResponse} from 'axios';
 import localStorage from '../libs/async-storage';
 import {TokenKeys} from '../libs/async-storage/constants/keys';
 
@@ -69,3 +69,48 @@ const onRejected = async (error: AxiosError) => {
 };
 
 axiosApi.interceptors.response.use(onFulfilled, onRejected);
+
+export interface BaseResponse<T> {
+  success: boolean;
+  errorCode: string | null;
+  message: string;
+  result: T;
+}
+
+export const GET = async <T>(
+  url: string,
+  config?: AxiosRequestConfig,
+): Promise<AxiosResponse<BaseResponse<T>>> => {
+  return axiosApi.get(url, config);
+};
+
+export const POST = async <T>(
+  url: string,
+  data?: any,
+  config?: AxiosRequestConfig,
+): Promise<AxiosResponse<BaseResponse<T>>> => {
+  return axiosApi.post(url, data, config);
+};
+
+export const DELETE = async <T>(
+  url: string,
+  config?: AxiosRequestConfig,
+): Promise<AxiosResponse<BaseResponse<T>>> => {
+  return axiosApi.delete(url, config);
+};
+
+export const PUT = async <T>(
+  url: string,
+  data?: any,
+  config?: AxiosRequestConfig,
+): Promise<AxiosResponse<BaseResponse<T>>> => {
+  return axiosApi.put(url, data, config);
+};
+
+export const PATCH = async <T>(
+  url: string,
+  data?: any,
+  config?: AxiosRequestConfig,
+): Promise<AxiosResponse<BaseResponse<T>>> => {
+  return axiosApi.patch(url, data, config);
+};
